+++ packages/cli/src/hooks/createTemplate.js
@@ -1,0 +1,101 @@
+
+const path = require('path');
+const inquirer = require('inquirer');
+const fsExtra = require('fs-extra');
+import { readRemoteFile } from './readFile'
+
+const currentDir = process.cwd();
+console.log('当前目录：', currentDir);
+
+
+/**
+ * 生成页面模版
+ */
+export const createPage = async function () {
+  // 输入页面名称
+  const { pageName } = await inquirer.prompt([
+    {
+      type: 'input',
+      name: 'pageName',
+      message: '请输入页面名称'
+    }
+  ]);
+
+  if (!pageName) {
+    console.error('Error: Page name is required.');
+    return;
+  }
+
+  const dirPath = path.join(currentDir, pageName);
+  console.log('创建目录：', dirPath);
+  // 确保目录不存在
+  if (await fsExtra.pathExists(dirPath)) {
+    console.error('Error: Page already exists.');
+    return;
+  }
+  // 创建文件夹
+  await fsExtra.ensureDir(dirPath);
+  const fs = require('fs');
+
+  try {
+    // 读取模板文件
+    const data = await readRemoteFile('https://raw.githubusercontent.com/wangchaoxx/projectTemplate/main/template/page/index.vue')
+    const indexData = await readRemoteFile('https://raw.githubusercontent.com/wangchaoxx/projectTemplate/main/template/page/index.data.ts')
+    // 创建文件
+    fs.writeFileSync(path.join(dirPath, 'index.vue'), data);
+    console.log('创建文件：', dirPath + '/index.vue');
+    fs.writeFileSync(path.join(dirPath, 'index.data.ts'), indexData);
+    console.log('创建文件：', dirPath + '/index.data.ts');
+    console.log(`页面创建成功！\n 请修改组件名称、路由名称、请求接口等信息`);
+    
+  } catch (error) {
+    // 下载模版文件错误，删除文件夹
+    fsExtra.removeSync(dirPath);
+  }
+}
+
+
+/**
+ * 生成弹框模版
+ */
+export const createModal = async () => {
+  
+  const { modalName } = await inquirer.prompt([
+    {
+      type: 'input',
+      name: 'modalName',
+      message: '请输入弹框名称'
+    }
+  ]);
+
+  if (!modalName) {
+    console.error('Error: Modal name is required.');
+    return;
+  }
+
+  const dirPath = path.join(currentDir, modalName);
+  // 确保目录不存在
+  if (fsExtra.pathExistsSync(dirPath)) {
+    console.error('Error: Modal already exists.');
+    return;
+  }
+  // 创建文件夹
+  fsExtra.ensureDirSync(dirPath);
+  console.log('创建目录：', dirPath);
+  const fs = require('fs');
+  // 读取模板文件
+  try {
+    const data = await readRemoteFile('https://raw.githubusercontent.com/wangchaoxx/projectTemplate/main/template/modal/index.vue')
+    const indexData = await readRemoteFile('https://raw.githubusercontent.com/wangchaoxx/projectTemplate/main/template/modal/index.data.ts')
+    // 创建文件
+    fs.writeFileSync(path.join(dirPath, 'index.vue'), data);
+    console.log('创建文件：', dirPath + '/index.vue');
+    fs.writeFileSync(path.join(dirPath, 'index.data.ts'), indexData);
+    console.log('创建文件：', dirPath + '/index.data.ts');
+    console.log(`弹框创建成功！\n 请修改组件名称、弹框标题、请求接口等信息`);
+  } catch (error) {
+    // 下载错误删除文件夹
+    fsExtra.removeSync(dirPath);
+  }
+
+}
